requirements.txt
Flask==3.0.0
Jinja2==3.1.6
Werkzeug==3.0.1
itsdangerous==2.2.0
click==8.1.7
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
pymongo==4.6.1
email-validator==2.0.0